% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.obligato_composition}
\alias{print.obligato_composition}
\title{Print an Obligato composition result}
\usage{
\method{print}{obligato_composition}(x, ...)
}
\arguments{
\item{x}{An \code{obligato_composition} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Obligato composition result
}
