% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.obligato_reffree}
\alias{print.obligato_reffree}
\title{Print an Obligato reference-free result}
\usage{
\method{print}{obligato_reffree}(x, ...)
}
\arguments{
\item{x}{An \code{obligato_reffree} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Obligato reference-free result
}
